refactor: move constraint conversion and printing functionality
