+++ src/pict-constraints-parser/index.ts
@@ -1,0 +1,1 @@
+export { printCodeFromAST } from './printer'
