+++ src/shared/helpers/utils.ts
@@ -1,0 +1,3 @@
+export function uuidv4(): string {
+  return crypto.randomUUID()
+}
